Created separate fluent navigation test file
Tried adding fluent tests inside existing navigation tests, like
'Navigate Link', but the asserts don't match because oldState isn't
populated in fluent navigation.
Fixing up Navigation tests so there's only one 'it' with multiple
asserts instead of multiple 'its' with one assert.
